requirements.txt
fastapi==0.68.1
python-multipart
uvicorn
moviepy
vosk
numpy
Pillow
python-dotenv